updated docs, shell integration, --repl

+++ commander.sh
@@ -1,24 +1,56 @@
+#Source this from your ~/.bashrc or ~/.zshrc
+#customize as you see fit
+#For example, you may want to integrate virtualenvwrapper
+
 commander-on() {
-    export COMMANDER_FIFO="/tmp/commander_$$.fifo"
-    [ -p "${COMMANDER_FIFO}" ] || mkfifo "${COMMANDER_FIFO}"
-    workon commander
-    unset RPROMPT
-    PS1="%~->"
-    ~/projects/unix_shell/commander.py --in "${COMMANDER_FIFO}" &
-    export COMMANDER_PID=$!
-    echo commander running with pid "${COMMANDER_PID}"
+    #Set this to "fifo" to have 1 long-running commander process
+    export COMMANDER_MODE=one
+    case "${COMMANDER_MODE}" in
+        "fifo")
+            export COMMANDER_FIFO="/tmp/commander_$$.fifo"
+            [ -p "${COMMANDER_FIFO}" ] || mkfifo "${COMMANDER_FIFO}"
+            workon commander
+            ~/dotfiles/commander.py --repl --in "${COMMANDER_FIFO}" &
+            export COMMANDER_PID=$!
+            echo commander running with pid "${COMMANDER_PID}"
+        ;;
+        "one")
+            unset COMMANDER_FIFO
+            unset COMMANDER_PID
+        ;;
+    esac
 }
 
 commander-off() {
-    [ -d "/proc/${COMMANDER_PID}" ] && kill "${COMMANDER_PID}"
+    [ -n "${COMMANDER_PID}" ] && [ -d "/proc/${COMMANDER_PID}" ] && kill "${COMMANDER_PID}"
     unset COMMANDER_PID
     [ -p "${COMMANDER_FIFO}" ] && rm "${COMMANDER_FIFO}"
     unset COMMANDER_FIFO
-    deactivate
-    #unfunction command_not_found_handler
+    unset COMMANDER_MODE
 }
 
-command_not_found_handler() {
-    [ -p "${COMMANDER_FIFO}" ] && echo "${@}" > "${COMMANDER_FIFO}"
+
+commander-exec() {
+    case "${COMMANDER_MODE}" in
+        fifo)
+            clear
+            echo "${@}" > "${COMMANDER_FIFO}"
+        ;;
+        one)
+             ~/dotfiles/commander.py "${@}"
+        ;;
+    esac
     return $?
 }
+
+if [ -n "${ZSH_VERSION} "]; then
+    #ZSH
+    command_not_found_handler() {
+        commander-exec "${@}"
+    }
+else
+    #BASH
+    command_not_found_handle() {
+        commander-exec "${@}"
+    }
+fi
